requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
jose[cryptography]
passlib[bcrypt]
pydantic
pydantic-settings
email-validator
python-dotenv
gunicorn